Resolve Relation frame interval problem.
Now Relations can be frame-less or frame-defined.
Relations are frame-less if when add_relation is called, no frame_intervals are defined. This
expresses the intention of the user to declare such Relation as non-defined at specific frames,
so it will NOT appear inside frame containers. This can be useful for causality relations.

Relations are frame-specific if when add_relation is called, frame_intervals are defined. Relations appear
in frames and stringify_frame calls.
update_relation can be called to update the frame_intervals of Relations.

RDFs frame_intervals are no longer modifying Relation frame intervals, as some Relations might need union,
others intersection, and others are better represented frame-less.

Now the user is responsible for defining the frame information of Relations.